test-django-asgi-upload/requirements.txt
Django

channels
daphne

psycopg  # connection to Postgres database
django-debug-toolbar
ipdb
strawberry-graphql
strawberry-graphql-django

-e ../../code/sentry-python  # sdk in a sibling folder to this projects folder
# sentry-sdk==1.34.0